refactor: вынесены типы, перемещён роутер, приведён код к единому стилю
