+++ src/components/map/map.tsx
@@ -2,7 +2,7 @@ import { useEffect, useRef } from 'react';
 import leaflet, { layerGroup, Marker } from 'leaflet';
 import 'leaflet/dist/leaflet.css';
 import useMap from '../../hooks/use-map';
-import { OfferDto } from '../../types/types';
+import { OfferDto } from '../../types/offer-dto';
 import { URL_MARKER_CURRENT, URL_MARKER_DEFAULT } from '../../const';
 import clsx from 'clsx';
