+++ src/pages/offer-screen/components/offer-host.tsx
@@ -11,6 +11,7 @@ interface Props {
 
 function OfferHost({ description, host }: Props) {
   const { name, avatarUrl, isPro } = host;
+
   return (
     <div className="offer__host">
       <h2 className="offer__host-title">Meet the host</h2>
