+++ src/services/api.ts
@@ -3,7 +3,6 @@ import { getToken } from './token';
 import { StatusCodes } from 'http-status-codes';
 import { toast } from 'react-toastify';
 
-
 type DetailMessageType = {
   type: string;
   message: string;
