+++ src/store/index.ts
@@ -29,5 +29,5 @@ export const store = configureStore({
 export const useAppDispatch = () => useDispatch<AppDispatch>();
 
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
-export type { RootState };
 
+export type { RootState };
