+++ src/types/comment-dto.ts
@@ -1,0 +1,11 @@
+export interface CommentDto {
+  id: string;
+  date: string;
+  user: {
+    name: string;
+    avatarUrl: string;
+    isPro: boolean;
+  };
+  comment: string;
+  rating: number;
+}
