+++ src/types/offer-dto.ts
@@ -1,15 +1,3 @@
-export interface CommentDto {
-  id: string;
-  date: string;
-  user: {
-    name: string;
-    avatarUrl: string;
-    isPro: boolean;
-  };
-  comment: string;
-  rating: number;
-}
-
 export interface OfferDetailsDto {
   id: string;
   title: string;
