+++ src/types/state.ts
@@ -1,6 +1,7 @@
 import { AuthorizationStatus } from '../const';
 import { store } from '../store';
-import { CommentDto, OfferDetailsDto, OfferDto } from './types';
+import { CommentDto } from './comment-dto';
+import { OfferDetailsDto, OfferDto } from './offer-dto';
 import { UserData } from './user-data';
 
 export interface ReviewsState {
